CMOS-100: split CI between PR and nightly + add extra smoke test
This commit splits the CI test suite into two: a set that runs on each PR, which only runs the smoke tests on containers (to get as fast a turnaround as possible), and a separate action which runs the full suite - run overnight by cron, or can also be triggered manually.

Also, buff up the smoke suite a bit, since that's all we'll run.