CMOS-112: Switch to streaming for cluster information
Previously we would get cluster node updates by polling /pools/default/nodeServices. This approach has a couple flaws described in CMOS-112, namely that it only includes nodes that haven't been failed over (which we may want to monitor, if only for the sake of alerting that they've been failed over). Also, polling is not the most efficient way of getting updates, and could lead to us missing quick cluster topology changes.

This commit changes the cluster monitoring subsystem to use the /poolsStreaming/default endpoint, which uses a long-running HTTP connection to stream changes to the /pools/default data. This should mean that we get updates almost instantly if nodes in the cluster change their membership. When the node we're streaming from itself leaves the cluster, it will close the connection - we'll handle this by switching to another node from its previous message. (We should have automated tests to verify this, but I have checked it manually.)

We use an exponential back-off algorithm in the event of streaming failures to avoid overloading ns_server.

I've tested this manually - we should add automatic tests, perhaps as part of the smoke suite or container integration (to simulate node failures and the like), as well as unit tests to the underlying functionality. I'll capture a subtask on the CMOS-104 epic.