CMOS-118, CMOS-119: Fix single-cluster overview dashboard
Due to a Grafana bug (https://github.com/grafana/grafana/issues/4234 / https://github.com/grafana/grafana/issues/40634), we can't use Regex escaping if cluster names can contain special characters. To resolve this, disable multi-value support for the $cluster variable and switch all queries to use doublequote instead.

Also change the cluster/node UUID variable sources to refer to cbmultimanager's Prometheus metrics, instead of the Couchbase Exporter or the JSON API. This adds more resilience in the face of a broken or misconfigured exporter. In theory this is still vulnerable to cbmultimanager or Prometheus as a whole going down, but if that happens CMOS as a whole is severely degraded and I'm not sure there's much sensible we can do then.