config-svc: disable cgo for microlith build (#54)
* config-svc: disable cgo for microlith build

For consistency, we would use cgo to build config-svc, but that makes portability difficult since it has to link against various system libraries, plus it can cause build issues. We don't actually need it though (unlike e.g. cbmultimanager which uses sqlite)

Disable cgo and set the `netgo` build tag to use a pure-go version of the net package.

Tested a `make container`.

* Remove unneeded linker flags

This should be a static-enough build to run (and I have tested running it in the Alpine build)