CMOS-131 Remove cluster name label
The `cluster` label on the `couchbase-server` jobs is unnecessary - the Exporter sets a label `cluster` and cbmultimanager sets `cluster_name` and `cluster_uuid`. If we set one at the job level it'll get clobbered or renamed to `exported_cluster` which may break the dashboards.

Remove the field entirely to avoid the need to keep around various versions of it. The generated Prometheus targets file will be named after the first node in the cluster (sanitised for a file name).

We may need to revisit this when we move to CB7's Prometheus, as I don't think it sets a `cluster` label - but for now this should be sufficient.