Allow changing docs contents for the packaged environment
Some parts of the docs only make sense when packaged inside the container, while others only make sense outside it. This commit introduces the `env-packaged` attribute, which can be used in an AsciiDoc `ifdef` to have conditional content.

For example, when built inside the container, the Support page links directly to the collect info form, while when built outside it instead has the URL of the form.