CMOS-179: Disable Promtail by default as a workaround for crash issues
In some cases we have seen Promtail crash immediately upon startup - see CMOS-179 for details. This appears to be either machine-specific, or related to some configuration.

Disable Promtail by default as a workaround. We don't actually have any dashboards that make use of Promtail's logging, so we don't lose anything.

I'll convert CMOS-179 into a tracking issue for re-enabling it.