Rework alert rule includes for Cluster Monitor changes
We moved the Cluster Monitor's alerting rules into its own repository,
so that the rules can be in sync with changes there, making it easier
to have bespoke, descriptive rules for each checker.

This commit reworks the build process to copy those rules over. It
also makes minor changes to the Alertmanager configuration to include
the "Remediation" field that those rules have.