CMOS-6: Add notices for 3rd party components
Pull in the notices files that Black Duck generates at build-time and include them in the built image. This does create an effect where build N will have the notices for build N-1 - but the notices shouldn't be changing that often towards the end of a release cycle anyway.

Also update the licensing docs to refer to them.