CMOS-250: Add unit tests for Loki rules
Add a framework for importing a set of logs into Loki (using Fluent Bit), and then asserting that the relevant alerting rule's query matches. Also add a GitHub Action to run it.

Note that this won't test the labels sent to Alertmanager (as that'd be too slow), only whether the query matches.