Dashboard polish (#237)
* Cluster/Node Overview: fix "Memory Usage" graph
We used to calculate it as mem_free / mem_total, which is semantically the opposite of "memory usage". Change it to mem_actual_used / mem_total.

* Data Service: tweak read/write failures
Change them to show the increase over the past 24 hours and clarify this in the title. Also, move them lower down, as they're a relatively unimportant stat (assuming things are working fine).

* Inventory, Cluster Overview: add feedback link